exchange float and int to int and float

+++ python_variable_annotations/6-sum_mixed_list.py
@@ -6,6 +6,6 @@ of integers and floats and returns their sum as a float
 import typing
 
 
-def sum_mixed_list(mxd_lst: typing.List[typing.Union[float, int]]) -> float:
+def sum_mixed_list(mxd_lst: typing.List[typing.Union[int, float]]) -> float:
     '''add a list of mixed float and int'''
     return float(sum(mxd_lst))
